Refactor marker clusters into single module
